+++ src/main/java/com/crud/crudstoreserver/models/Product.java
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 @Data
 @Entity
@@ -15,7 +16,7 @@ public class Product {
     private Long id;
 
     private String name;
-    private int price;
+    private BigDecimal price;
     private String description;
     private ProductType productType;
     private int size;
